Update
Added 99 blocks with texture of cobblestone.
Localization of blocks has been carried out: Cobblestone into Russian.
Fixed block numbering for Russification.
The texture of the 4rd block has been fixed.
The texture of the 6rd block has been fixed.
The texture of the 47rd block has been fixed.
The texture of the 91rd block has been fixed.
The texture of the 96rd block has been fixed.
The texture of the 97rd block has been fixed.
The texture of the 15rd block has been fixed.
The texture of the 71rd block has been fixed.
The texture of the 69rd block has been fixed.
The texture of the 29rd block has been fixed.
The texture of the 31rd block has been fixed.
The texture of the 41rd block has been fixed.
The texture of the 57rd block has been fixed.